dijkstra_drones/requirements.txt
Django>=4.0
djangorestframework
networkx
leaflet